backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "channels>=4.3.1",
    "channels-redis>=4.3.0",
    "django>=5.2.6",
    "django-cors-headers>=4.9.0",
    "django-extensions>=4.1",
    "django-ratelimit>=4.1.0",
    "djangorestframework>=3.16.1",
    "dotenv>=0.9.9",
    "numpy>=2.3.3",
    "openai>=1.109.1",
    "pandas>=2.3.2",
    "pdfplumber>=0.11.7",
    "psycopg2-binary>=2.9.10",
    "pymupdf>=1.26.4",
    "pytest-env>=1.1.5",
    "python-dotenv>=1.1.1",
    "python-magic>=0.4.27",
    "redis>=6.4.0",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.1",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "isort>=6.0.1",
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
